Grouped AssetBundles to one folder
