+++ src/Asset/UserswitchAsset.php
@@ -1,0 +1,12 @@
+<?php
+namespace Yiisoft\Yii\Debug\Asset;
+
+use Yiisoft\Assets\AssetBundle;
+
+class UserswitchAsset extends AssetBundle
+{
+    public ?string $sourcePath = '@Yiisoft/Yii/Debug/assets';
+    public array $js = [
+        'js/userswitch.js',
+    ];
+}
